src/main/java/Radio.java
public class Radio {
    private int currentRadioStation;
    private int currentSoundVolume;

    // RADIO STATION
    public int getCurrentRadioStation() {
        return currentRadioStation;
    }

    public void setCurrentRadioStation(int currentRadioStation) {
        if (currentRadioStation < 0) {
            return;
        }
        if (currentRadioStation > 9) {
            return;
        }
        this.currentRadioStation = currentRadioStation;
    }

    public void next() {
        if (currentRadioStation != 9) {
            currentRadioStation++;
        } else {
            currentRadioStation = 0;
        }
    }

    public void prev() {
        if (currentRadioStation != 0) {
            currentRadioStation--;
        } else {
            currentRadioStation = 9;
        }
    }

    // SOUND VOLUME
    public int getCurrentSoundVolume() {
        return currentSoundVolume;
    }

    public void setCurrentSoundVolume(int currentSoundVolume) {
        if (currentSoundVolume < 0) {
            return;
        }
        if (currentSoundVolume > 10) {
            return;
        }
        this.currentSoundVolume = currentSoundVolume;
    }

    public void increaseVolume() {
        if (currentSoundVolume != 10) {
            currentSoundVolume++;
        } else {
            currentSoundVolume = 10;
        }
    }

    public void decreaseVolume() {
        if (currentSoundVolume != 0) {
            currentSoundVolume--;
        } else {
            currentSoundVolume = 0;
        }
    }
}
